+++ docs/releases.md
@@ -1,3 +1,6 @@
+# Unreleased
+- Legs can be drawn using only their vertical change in the extended elevation (thanks Andrew Atkinson)
+
 # 2026-07-08 1.12.3
 - Fix Survex and Therion backwards leg bug (thanks Andrew Atkinson)